+++ src/Farmer/Builders/Builders.ContainerService.fs
@@ -4,6 +4,7 @@ module Farmer.Builders.ContainerService
 open System
 open Farmer
 open Farmer.Arm
+open Farmer.Arm.RoleAssignment
 open Farmer.Identity
 open Farmer.Vm
 
@@ -48,9 +49,12 @@ type NetworkProfileConfig =
 type AksConfig =
     { Name : ResourceName
       AgentPools : AgentPoolConfig list
+      Dependencies : ResourceId Set
+      DependencyExpressions : ArmExpression Set
       DnsPrefix : string
       EnableRBAC : bool
       Identity : ManagedIdentity
+      IdentityProfile : ManagedClusterIdentityProfile option
       ApiServerAccessProfile : ApiServerAccessProfileConfig option
       LinuxProfile : (string * string list) option
       NetworkProfile : NetworkProfileConfig option
@@ -61,15 +65,17 @@ type AksConfig =
     interface IBuilder with
         member this.ResourceId = this.ResourceId
         member this.BuildResources location = [
-            // VM itself
             { Name = this.Name
               Location = location
+              Dependencies = this.Dependencies
+              DependencyExpressions = this.DependencyExpressions
               DnsPrefix =
                   if String.IsNullOrWhiteSpace this.DnsPrefix then
                       $"{this.Name.Value}-%x{this.Name.Value.GetHashCode()}"
                   else this.DnsPrefix
               EnableRBAC = this.EnableRBAC
               Identity = this.Identity
+              IdentityProfile = this.IdentityProfile
               AgentPoolProfiles =
                 match this.AgentPools with
                 | [] ->
@@ -221,10 +227,13 @@ let private (|PrivateClusterEnabled|_|) =
 type AksBuilder() =
     member _.Yield _ =
         { Name = ResourceName.Empty
+          Dependencies = Set.empty
+          DependencyExpressions = Set.empty
           AgentPools = []
           DnsPrefix = ""
           EnableRBAC = false
           Identity = ManagedIdentity.Empty
+          IdentityProfile = None
           ApiServerAccessProfile = None
           LinuxProfile = None
           NetworkProfile = None
@@ -249,6 +258,10 @@ type AksBuilder() =
     member _.EnableRBAC(state:AksConfig) = { state with EnableRBAC = true }
     /// Sets the managed identity on this cluster.
     interface IIdentity<AksConfig> with member _.Add state updater = { state with Identity = updater state.Identity }
+    /// Support for "depends_on"
+    interface IDependable<AksConfig> with member _.Add state newDeps = { state with Dependencies = state.Dependencies + newDeps }
+    [<CustomOperation "depends_on_expression">]
+    member _.DependencyExpressions(state:AksConfig, dependencyExpr:ArmExpression) = { state with DependencyExpressions = state.DependencyExpressions.Add dependencyExpr }
     /// Adds agent pools to the AKS cluster.
     [<CustomOperation "add_agent_pools">]
     member _.AddAgentPools(state:AksConfig, pools) = { state with AgentPools = state.AgentPools @ pools }
@@ -271,6 +284,14 @@ type AksBuilder() =
             | None -> { AuthorizedIPRanges = range; EnablePrivateCluster = None }
             | Some profile -> { profile with AuthorizedIPRanges = profile.AuthorizedIPRanges @ range }
         { state with ApiServerAccessProfile = Some accessProfile }
+    /// Sets the kubelet identity for managing access to an Azure Container Registry
+    [<CustomOperation "kubelet_identity">]
+    member _.KubeletIdentity (state:AksConfig, identity:ResourceId) =
+        match state.IdentityProfile with
+        | None -> { state with IdentityProfile = Some { KubeletIdentity = Some identity }  }
+        | Some identityProfile -> { state with IdentityProfile = Some { identityProfile with KubeletIdentity = Some identity }  }
+    member this.KubeletIdentity (state:AksConfig, identity:UserAssignedIdentity.UserAssignedIdentityConfig) =
+        this.KubeletIdentity (state, identity.ResourceId)
     /// Sets the network profile for the AKS cluster.
     [<CustomOperation "network_profile">]
     member _.NetworkProfile(state:AksConfig, networkProfile) = { state with NetworkProfile = Some networkProfile }